Mesh Generation Refinement
First Stage of refinement of the mesh generator. Edited code to make it easier to understand and also linked the boundary squares together to make the terrain a little smoother. Still requires more refinement between the boundary squares (such as sub-nodes to allow for smaller triangles therefore smoother looking terrain) as it is still extremely bumpy.