Worked on Curved terrain
Had to rewrite the vertices and triangle assignment methods to allow for the curves to be implemented properly as meshes. Also rewrote a seperate method for creating meshes from a single square making it easier to create the flat terrain. Currently have the curves implemented as quadratic curves but need to add functionality for cubic curves to allow for smoother curve transition between the squares.