Added Primitive health system
Added lives to the car and set the lives to subtract every time the player hits an obstacle. Also added a invincibility on damage taken, so the player will be invincible for 3 seconds after taking damage. I plan to make the sprite flash on damage too, and add HUD elements to reflect health of the player