Added Comments & Fixed a bug
Added comments to methods to describe what they do. Also fixed a bug with the collision points when accessing the connecting square between two sections. (It used the connecting squares x value which was the the value of (width - 1) causing the collision points to bug out in the new section.) Setting the connecting squares x values to 0 fixed this.