Attempted a fix for the Read/Write Issue, kinda got somewhere.
The SaveState.txt DOES change values but only after some sort of event of which I'm not too sure. One time it's the same then other times all of a sudden it's changed to the correct expected values. Also refined the UpgradeMenu script so it's easy on the eyes. Well my eyes.  Appropriately edited the event triggers for each button in the Upgrades Menu and also reordered the values of the text file so it's Coins, Health and Speed instead of Health, Coins and Speed. And stop asking about the TestScript stuff. It's for me and doesn't concern you.