Variety of changes and adjustments
Rewrote the upgrades script making it easier to add new upgrades during development of the game.

Added lifetime distance counter which Jadan started, now fully works and stores in database.

Fixed car image so the background is clear.

Adjusted the jump script to allow for the same X velocity when they jump. Previously it set it to a fixed X velocity causing sudden stops when going really fast.

Adjusted generation of clouds, they now generate at a random height in a fixed range.